requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
mysql-connector-python
pymysql
sqlalchemy
tiktoken